fix(economy):  Move deposit, withdraw commands to moneyManagement/
