+++ commands/economy/economy.go
@@ -4,8 +4,8 @@ package economy
 
 import (
 	"github.com/RhykerWells/asbwig/commands/economy/informational/balance"
-	"github.com/RhykerWells/asbwig/commands/economy/moneyMaking/deposit"
-	"github.com/RhykerWells/asbwig/commands/economy/moneyMaking/withdraw"
+	"github.com/RhykerWells/asbwig/commands/economy/moneyManagement/deposit"
+	"github.com/RhykerWells/asbwig/commands/economy/moneyManagement/withdraw"
 	"github.com/RhykerWells/asbwig/commands/economy/settings/set"
 	"github.com/RhykerWells/asbwig/commands/economy/settings/viewsettings"
 	"github.com/RhykerWells/asbwig/common"
